passport

+++ controllers/resController.js
@@ -16,24 +16,44 @@ export const postResUpload = async (req, res) => {
         body : {title, description},
         file : {path}
     } = req;
-    const restaurants = await resModel.create({
-        photo: path,
-        title,
-        description
-    })
-    res.redirect(routes.resDetail(restaurants.id));
+    try {
+        const restaurants = await resModel.create({
+            photo: path,
+            title,
+            description,
+            creator: req.user.id
+        });
+        req.user.restaurants.push(restaurants._id);
+        req.user.save();
+        res.redirect(routes.resDetail(restaurants.id));
+    } catch (error) {
+        console.log(error);
+        res.redirect(routes.home);
+    }
+    
 }
 
 export const resDetail = async (req, res) => {
     const {params:{id}} = req;
-    const restaurants = await resModel.findById(id)
-    res.render("resDetail", { pageName: "ResDetail", restaurants });
+    try {
+        const restaurants = await resModel.findById(id).populate("creator")
+        res.render("resDetail", { pageName: "ResDetail", restaurants });
+    } catch (error) {
+        console.log(error);
+        res.redirect(routes.home)
+    }
+    
 }
 
 export const getResEdit = async (req, res) => {
     const {params:{id}} = req;
-    const restaurants = await resModel.findById(id)
-    res.render("resEdit", { pageName: "RedEdit", restaurants });
+    try {
+        const restaurants = await resModel.findById(id)
+        res.render("resEdit", { pageName: "RedEdit", restaurants });
+    } catch (error) {
+        console.log(error);
+        res.redirect(routes.home);
+    }
 }
 
 export const postResEdit = async (req, res) => {
@@ -41,20 +61,46 @@ export const postResEdit = async (req, res) => {
         params:{id},
         body: {title, description}
     } = req;
-    await resModel.findByIdAndUpdate({_id : id}, {title, description})
-    res.redirect(routes.resDetail(id));
+    try {
+        const restaurant = await resModel.findById(id).populate("creator");
+        if ( restaurant.creator.id === req.user.id ) {
+            await resModel.findByIdAndUpdate({_id : id}, {title, description})
+            res.redirect(routes.resDetail(id));
+        } else {
+            res.redirect(routes.resDetail(id));
+        }   
+    } catch (error) {
+        console.log(error);
+        res.redirect(routes.resDetail(id));
+    }
 }
 
 export const resDelete = async (req, res) => {
     const {
         params:{id}
     } = req;
-    await resModel.findByIdAndDelete({_id : id});
-    res.redirect(routes.home);
+    try {
+        const restaurant = await resModel.findById(id).populate("creator");
+        if ( restaurant.creator.id === req.user.id ) {
+            await resModel.findByIdAndDelete({_id : id});
+            res.redirect(routes.home);
+        } else {
+            res.redirect(routes.home);
+        }
+    } catch (error) {
+        console.log(error);
+        res.redirect(routes.home);
+    }
 }
 
 export const search = async (req, res) => {
     const {query:{search}} = req;
-    const restaurants = await resModel.find({title:{ $regex : search , $options: 'i' }})
-    res.render("search", { pageName: "Search", search, restaurants });
+    try {
+        const restaurants = await resModel.find({title:{ $regex : search , $options: 'i' }})
+        res.render("search", { pageName: "Search", search, restaurants });
+    } catch (error) {
+        console.log(error);
+        res.redirect(routes.home);
+    }
+    
 }
